added link back to home on navbar
